requirement.txt
pandas
requests
apache-airflow[postgres]
sqlalchemy
pyarrow
python-dotenv
great_expectations
psycopg2


